requirements.txt
streamlit
pandas
sqlalchemy
pymysql
python-dotenv
openpyxl
fuzzywuzzy
